SW-26587 - Fix PHPStan issues with PHP 8 in components classes

See merge request shopware/5/product/shopware!767

+++ engine/Shopware/Components/CacheManager.php
@@ -265,7 +265,7 @@ class CacheManager
         $configuration = $this->emConfig;
 
         $metaDataCache = $configuration->getMetadataCacheImpl();
-        if (method_exists($metaDataCache, 'deleteAll')) {
+        if ($metaDataCache !== null && method_exists($metaDataCache, 'deleteAll')) {
             $metaDataCache->deleteAll();
         }
